+++ woocommerce/global/quantity-input.php
@@ -6,7 +6,7 @@
  * @subpackage Secretum
  * @author     SecretumTheme <author@secretumtheme.com>
  * @copyright  2018-2019 Secretum
- * @version    3.6.0
+ * @version    4.0.0
  * @license    https://github.com/SecretumTheme/secretum/blob/master/license.txt GPL-2.0
  * @link       https://github.com/SecretumTheme/secretum/blob/master/woocommerce/global/quantity-input.php
  * @since      1.0.0
@@ -25,10 +25,11 @@ if ( $max_value && $min_value === $max_value ) { ?>
 	<?php
 } else {
 	/* Translators: %s number of items in a shopping cart */
-	$secretum_labelledby = ! empty( $args['product_name'] ) ? sprintf( __( '%s quantity', 'secretum' ), wp_strip_all_tags( $args['product_name'] ) ) : '';
+	$secretum_label = ! empty( $args['product_name'] ) ? sprintf( __( '%s Quantity', 'secretum' ), wp_strip_all_tags( $args['product_name'] ) ) : esc_html__( 'Quantity', 'secretum' );
 	?>
 	<div class="quantity form-group row mb-3">
-		<label class="screen-reader-text" for="<?php echo esc_attr( $input_id ); ?>"><?php esc_html_e( 'Quantity', 'secretum' ); ?></label>
+		<?php do_action( 'woocommerce_before_quantity_input_field' ); ?>
+		<label class="screen-reader-text" for="<?php echo esc_attr( $input_id ); ?>"><?php $secretum_label; ?></label>
 		<input
 			type="number"
 			id="<?php echo esc_attr( $input_id ); ?>"
@@ -40,8 +41,10 @@ if ( $max_value && $min_value === $max_value ) { ?>
 			value="<?php echo esc_attr( $input_value ); ?>"
 			title="<?php echo esc_html_x( 'Qty', 'Product quantity input tooltip', 'secretum' ); ?>"
 			size="4"
+			placeholder="<?php echo esc_attr( $placeholder ); ?>"
 			inputmode="<?php echo esc_attr( $inputmode ); ?>"
 		/>
+		<?php do_action( 'woocommerce_after_quantity_input_field' ); ?>
 	</div>
 	<?php
 }